requirements.txt
streamlit
numpy
tensorflow
pillow
matplotlib
